Tied OutputGenerator into 'Simulation'
DREAM now uses the new OutputGenerator classes to save output. We still
need to give the solvers their own OutputGenerators though, and
implement support for saving the current iteration in the
OutputGenerators.
