+++ iface/Main.cpp
@@ -35,11 +35,8 @@ struct cmd_args {
     bool display_settings=false;
     bool print_adas=false;
     bool splash=true;
-    bool save_initial = false;
     string
-        input_filename,
-        initial_filename,
-        output_filename;
+        input_filename;
 };
 
 void display_settings(DREAM::Settings *s=nullptr) {
@@ -67,9 +64,7 @@ void print_help() {
     cout << "OPTIONS" << endl;
     cout << "  -a           Print list of elements in ADAS database." << endl;
     cout << "  -h           Print this help." << endl;
-    cout << "  -i           Save the initial simulation state to the named file." << endl;
     cout << "  -l           List all available settings in DREAM." << endl;
-    cout << "  -o           Specify the name of the output file." << endl;
     cout << "  -s           Do not show the splash screen." << endl;
 }
 
@@ -83,11 +78,9 @@ struct cmd_args *parse_args(int argc, char *argv[]) {
     char c;
 
     struct cmd_args *a = new struct cmd_args;
-    a->initial_filename = "";
-    a->output_filename = "";
     a->display_settings = false;
 
-    while ((c = getopt(argc, argv, "ahi:lo:s")) != -1) {
+    while ((c = getopt(argc, argv, "ahi:ls")) != -1) {
         switch (c) {
             case 'a':
                 a->print_adas = true;
@@ -95,15 +88,8 @@ struct cmd_args *parse_args(int argc, char *argv[]) {
             case 'h':
                 print_help();
                 break;
-            case 'i':
-                a->save_initial = true;
-                a->initial_filename = string(optarg);
-                break;
             case 'l':
                 display_settings();
-                break;
-            case 'o':
-                a->output_filename = string(optarg);
                 break;
             case 's':
                 a->splash = false;
@@ -230,9 +216,6 @@ int main(int argc, char *argv[]) {
         if (a->print_adas)
             display_adas(sim);
 
-        if (a->save_initial)
-            sim->Save(a->initial_filename);
-
         sim->Run();
     } catch (DREAM::FVM::FVMException &ex) {
         DREAM::IO::PrintError(ex.what());
@@ -247,10 +230,7 @@ int main(int argc, char *argv[]) {
 
     if (sim != nullptr) {
         try {
-            if (a->output_filename != "")
-                sim->Save(a->output_filename);
-            else
-                sim->Save();
+            sim->Save();
         } catch (H5::FileIException &ex) {
             DREAM::IO::PrintError(ex.getDetailMsg().c_str());
             exit_code = 4;
