+++ include/DREAM/EquationSystem.hpp
@@ -102,6 +102,8 @@ namespace DREAM {
         UnknownQuantityEquation *GetEquation(const len_t i) { return unknown_equations.at(i); }
         std::vector<UnknownQuantityEquation*> *GetEquations() { return &unknown_equations; }
 
+        std::vector<real_t>& GetTimes() { return this->times; }
+
         real_t *GetUnknownData(const len_t i) { return unknowns.GetUnknownData(i); }
         len_t GetUnknownID(const std::string& name) { return unknowns.GetUnknownID(name); }
         len_t GetNUnknowns() const { return this->unknowns.GetNUnknowns(); }
@@ -161,12 +163,7 @@ namespace DREAM {
         void SetSolver(Solver*);
         void SetTimeStepper(TimeStepper *ts) { this->timestepper = ts; }
 
-        void SaveGrids(SFile*, const std::string& path="");
-        void SaveIonMetaData(SFile*, const std::string& path="");
-        void SaveMomentumGrid(SFile*, const std::string&, FVM::Grid*, enum OptionConstants::momentumgrid_type);
         void SaveTimings(SFile*, const std::string&);
-        void WriteCopyArray(SFile*, const std::string&, const real_t *const*, const len_t, const len_t);
-        void WriteCopyMultiArray(SFile*, const std::string&, const real_t *const*, const sfilesize_t, const sfilesize_t[]);
 
         void Solve();
         void Rebuild();
