+++ src/OutputGeneratorSFile.cpp
@@ -3,25 +3,23 @@
  * SFile object.
  */
 
+#include <string>
 #include "DREAM/OutputGeneratorSFile.hpp"
 
 
 using namespace DREAM;
+using namespace std;
 
 /**
  * Constructor.
  */
 OutputGeneratorSFile::OutputGeneratorSFile(
-	FVM::Grid *grid, FVM::UnknownQuantityHandler *unknowns,
-	IonHandler *ions, OtherQuantityHandler *oqty,
 	EquationSystem *eqsys, const std::string& filename
-) : OutputGeneratorSFile(grid, unknowns, ions, oqty, eqsys, SFile::Create(filename, SFILE_MODE_WRITE)) {}
+) : OutputGeneratorSFile(eqsys, SFile::Create(filename, SFILE_MODE_WRITE)) {}
 
 OutputGeneratorSFile::OutputGeneratorSFile(
-	FVM::Grid *grid, FVM::UnknownQuantityHandler *unknowns,
-	IonHandler *ions, OtherQuantityHandler *oqty,
 	EquationSystem *eqsys, SFile *sf
-) : OutputGenerator(grid, unknowns, ions, oqty, eqsys), sf(sf) {}
+) : OutputGenerator(eqsys), sf(sf) {}
 
 /**
  * Destructor.
@@ -34,42 +32,213 @@ OutputGeneratorSFile::~OutputGeneratorSFile() {
 /**
  * Save grid data.
  */
-OutputGeneratorSFile::SaveGrids(const std::string& name) {
-    // Save grids
+void OutputGeneratorSFile::SaveGrids(const std::string& name) {
     this->sf->CreateStruct(name);
-    this->eqsys->SaveGrids(this->sf, name);
+
+    string group;
+    if (name.back() == '/')
+        group = name;
+    else
+        group = name + "/";
+
+    // Time grid
+    const real_t *t = this->eqsys->GetTimes().data();
+    this->sf->WriteList(group + "t", t, this->eqsys->GetTimes().size());
+
+    // Radial grid
+    const real_t *r   = this->fluidGrid->GetRadialGrid()->GetR();
+    const real_t *r_f = this->fluidGrid->GetRadialGrid()->GetR_f();
+    const real_t *dr  = this->fluidGrid->GetRadialGrid()->GetDr();
+    this->sf->WriteList(group + "r", r, this->fluidGrid->GetNr());
+    this->sf->WriteList(group + "r_f", r_f, this->fluidGrid->GetNr()+1);
+    this->sf->WriteList(group + "dr", dr, this->fluidGrid->GetNr());
+
+    // Volume elements
+    const real_t *VpVol = this->fluidGrid->GetVpVol();
+    this->sf->WriteList(group + "VpVol", VpVol, this->fluidGrid->GetNr());
+
+    // Hot-tail grid
+    if (this->hottailGrid != nullptr) {
+        this->sf->CreateStruct(group + "hottail");
+        SaveMomentumGrid(this->sf, group + "hottail/", this->hottailGrid, this->eqsys->GetHotTailGridType());
+    }
+
+    // Runaway grid
+    if (this->runawayGrid != nullptr) {
+        this->sf->CreateStruct(group + "runaway");
+        SaveMomentumGrid(this->sf, group + "runaway/", this->runawayGrid, this->eqsys->GetRunawayGridType());
+    }
 }
-    
+
+/**
+ * XXX Here we assume that all momentum grids are the same
+ * Save a momentum grid to the given SFile object.
+ *
+ * sf:       SFile object to write grids to.
+ * gridname: Full path to grid data in output file.
+ * g:        Grid to save.
+ */
+void OutputGeneratorSFile::SaveMomentumGrid(
+    SFile *sf, const string& gridname, FVM::Grid *g,
+    enum OptionConstants::momentumgrid_type tp
+) {
+    const real_t *p1   = g->GetMomentumGrid(0)->GetP1();
+    const real_t *p2   = g->GetMomentumGrid(0)->GetP2();
+    const real_t *p1_f = g->GetMomentumGrid(0)->GetP1_f();
+    const real_t *p2_f = g->GetMomentumGrid(0)->GetP2_f();
+    const real_t *dp1  = g->GetMomentumGrid(0)->GetDp1();
+    const real_t *dp2  = g->GetMomentumGrid(0)->GetDp2();
+    const len_t np1    = g->GetMomentumGrid(0)->GetNp1();
+    const len_t np2    = g->GetMomentumGrid(0)->GetNp2();
+    const len_t nr     = g->GetRadialGrid()->GetNr();
+
+    // Write grid type
+    sf->WriteInt32List(gridname + "type", (int32_t*)&tp, 1);
+
+    // Grid coordinates
+    sf->WriteList(gridname + "p1", p1, np1);
+    sf->WriteList(gridname + "p2", p2, np2);
+
+    // Flux grid coordinates
+    sf->WriteList(gridname + "p1_f", p1_f, np1+1);
+    sf->WriteList(gridname + "p2_f", p2_f, np2+1);
+
+    // Grid cell sizes
+    sf->WriteList(gridname + "dp1", dp1, np1);
+    sf->WriteList(gridname + "dp2", dp2, np2);
+
+    // Grid volumes
+    sfilesize_t dims[3] = {nr, np2, np1};
+    const real_t *const* Vp = g->GetVp();
+    WriteCopyMultiArray(sf, gridname + "Vprime", Vp, 3, dims);
+}
+
 /**
  * Save ion meta data.
  */
-OutputGeneratorSFile::SaveIonMetaData(const std::string& name) {
-    // Save ion metadata
+void OutputGeneratorSFile::SaveIonMetaData(const std::string& name) {
     this->sf->CreateStruct(name);
-    this->eqsys->SaveIonMetaData(this->sf, name);
+
+    string group;
+    if (name.back() == '/')
+        group = name;
+    else
+        group = name + "/";
+
+    // Get list of charges
+    const len_t nZ = this->ions->GetNZ();
+    const len_t *Z = this->ions->GetZs();
+    sf->WriteList(group + "Z", Z, nZ);
+
+    // Construct name vector
+    const vector<string> namelist = this->ions->GetNameList();
+
+    string names = "";
+    for (auto it = namelist.begin(); it != namelist.end(); it++)
+        names += *it + ";";
+
+    sf->WriteString(group + "names", names);
 }
 
 /**
  * Save other quantities.
  */
-OutputGeneratorSFile::SaveOtherQuantities(const std::string& name) {
+void OutputGeneratorSFile::SaveOtherQuantities(const std::string& name) {
+    this->sf->CreateStruct(name);
 	this->oqty->SaveSFile(sf, name);
+}
+
+/**
+ * Save settings used for this simulation
+ * TODO TODO TODO
+ */
+void OutputGeneratorSFile::SaveSettings(const std::string&) {
 }
 
 /**
  * Save timing information.
  */
-OutputGeneratorSFile::SaveTimings(const std::string& name) {
-    // Save timing information
+void OutputGeneratorSFile::SaveTimings(const std::string& name) {
     this->eqsys->SaveTimings(this->sf, name);
 }
 
 /**
  * Save unknown quantities.
  */
-OutputGeneratorSFile::SaveUnknowns(const std::string& name) {
-    // Save unknowns
+void OutputGeneratorSFile::SaveUnknowns(const std::string& name) {
     this->sf->CreateStruct(name);
     this->unknowns->SaveSFile(this->sf, name, false);
+}
+
+/**
+ * Save the given array to the specified SFile with the
+ * given name. The array is copied before writing so that
+ * 'v' need not be stored contiguously in memory.
+ *
+ * sf:   SFile object to write array to.
+ * name: Name of variable in file.
+ * v:    2D array data to write.
+ * m:    Length of first dimension of 'v'.
+ * n:    Length of second dimension of 'v'.
+ */
+void OutputGeneratorSFile::WriteCopyArray(
+    SFile *sf, const string& name, const real_t *const* v,
+    const len_t m, const len_t n
+) {
+    // First, copy array to make it contiguous
+    real_t **t = new real_t*[m];
+    t[0] = new real_t[m*n];
+
+    for (len_t i = 1; i < m; i++)
+        t[i] = t[i-1] + n;
+
+    for (len_t i = 0; i < m; i++)
+        for (len_t j = 0; j < n; j++)
+            t[i][j] = v[i][j];
+
+    sf->WriteArray(name, t, m, n);
+    
+    delete [] t[0];
+    delete [] t;
+}
+
+/**
+ * Save the given multi-dimensional array to the specified
+ * SFile with the given name. The array is copied before
+ * writing so that the first dimension of 'v' need not be
+ * stored contiguously in memory.
+ *
+ * sf:   SFile object to write array to.
+ * name: Name of variable in file.
+ * v:    Multi-dimensional array data to write.
+ * m:    Length of first dimension of 'v'.
+ * n:    Length of second dimension of 'v'.
+ */
+void OutputGeneratorSFile::WriteCopyMultiArray(
+    SFile *sf, const string& name, const real_t *const* v,
+    const sfilesize_t ndim, const sfilesize_t dims[]
+) {
+    if (ndim < 2)
+        throw OutputGeneratorException("Saving '%s': Invalid number of dimensions of multi-dimensional array. Unless this is a bug, 'SFile::WriteList()' should be used instead.", name.c_str());
+
+    // Calculate total number of elements in array
+    len_t totSize = 1;
+    for (len_t i = 0; i < ndim; i++)
+        totSize *= dims[i];
+
+    // Total size of all but the first dimension
+    // (e.g. of the momentum grid at each radius)
+    len_t otherSize = totSize / dims[0];
+
+    // Copy array to make it contiguous
+    real_t *t = new real_t[totSize];
+
+    for (len_t i = 0; i < dims[0]; i++)
+        for (len_t j = 0; j < otherSize; j++)
+            t[i*otherSize + j] = v[i][j];
+
+    sf->WriteMultiArray(name, t, ndim, dims);
+    
+    delete [] t;
 }
